requirements.txt
numpy
matplotlib
scipy
pyam-iamc
pandas